🎨 일정 페이지 2차 업데이트
디자인 수정 반영 + 필요없는 페이지 삭제
