+++ lib/main.dart
@@ -8,7 +8,6 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:taekwon/home/home_navigator.dart';
 import 'package:taekwon/mypage/mypage_screen.dart'; // mypage
 import 'package:taekwon/navigation/navigation_bar.dart';
-import 'package:taekwon/schedule/add_schedule.dart';
 import 'package:taekwon/schedule/schedule_screen.dart'; // navigation_bar
 
 void main() {
@@ -52,7 +51,7 @@ class _MainScaffoldState extends State<MainScaffold> {
   int _currentIndex = 0;
 
   final List<Widget> _screens = [
-    const AddSchedulePage(), //  포인트샵 화면...인데 확인차 로그인페이지를 넣어뒀어요!! 나중에 수정예정
+    const SchedulePage(), //  포인트샵 화면...인데 확인차 다른페이지를 넣어뒀어요!! 나중에 수정예정
     const SchedulePage(), //  일정 화면
     const HomeNavigator(), //  홈 화면
     const MyPage(), //  마이페이지 화면
