--- lib/schedule/add_schedule.dart
@@ -1,244 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:taekwon/decoration/color_palette.dart';
-
-import 'add_student.dart'; // 실제 경로에 맞게 수정
-
-class AddSchedulePage extends StatefulWidget {
-  const AddSchedulePage({super.key});
-
-  @override
-  State<AddSchedulePage> createState() => _AddSchedulePageState();
-}
-
-class _AddSchedulePageState extends State<AddSchedulePage> {
-  TimeOfDay? selectedTime;
-  int? _attendanceNumber; // 선택된 출석 인원 저장 변수
-
-  void _showTimePicker() {
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return Dialog(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(16),
-          ),
-          child: Container(
-            height: 300,
-            padding: const EdgeInsets.all(16),
-            color: Colors.white, // 다이얼로그 전체 배경 흰색
-            child: Column(
-              children: [
-                Expanded(
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.circular(12), // 타임피커 모서리 둥글게
-                    child: CupertinoTheme(
-                      data: const CupertinoThemeData(
-                        brightness: Brightness.light,
-                      ),
-                      child: CupertinoDatePicker(
-                        backgroundColor: Colors.white, // 타임피커 배경 흰색
-                        mode: CupertinoDatePickerMode.time,
-                        use24hFormat: false, // 오전/오후 표시
-                        initialDateTime: DateTime.now(),
-                        onDateTimeChanged: (DateTime value) {
-                          setState(() {
-                            selectedTime = TimeOfDay.fromDateTime(value);
-                          });
-                        },
-                      ),
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 8),
-                ElevatedButton(
-                  onPressed: () => Navigator.pop(context),
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: const Color(0xFF2962FF),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                  ),
-                  child: const Text(
-                    "완료",
-                    style: TextStyle(color: Colors.white),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildAttendanceField() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const Text("출석 인원", style: TextStyle(fontSize: 12, color: Colors.grey)),
-        const SizedBox(height: 4),
-        GestureDetector(
-          onTap: () async {
-            final result = await Navigator.push<int>(
-              context,
-              MaterialPageRoute(builder: (context) => const AddStudentPage()),
-            );
-            if (result != null) {
-              setState(() {
-                _attendanceNumber = result;
-              });
-            }
-          },
-          child: Container(
-            height: 48,
-            padding: const EdgeInsets.symmetric(horizontal: 12),
-            alignment: Alignment.centerLeft,
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.grey),
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Text(
-              _attendanceNumber != null ? '$_attendanceNumber 명 선택됨' : '선택하기',
-              style: const TextStyle(fontSize: 14, color: Colors.black87),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildLabeledField(String label) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(label, style: const TextStyle(fontSize: 12, color: Colors.grey)),
-        const SizedBox(height: 4),
-        TextField(
-          decoration: InputDecoration(
-            contentPadding: const EdgeInsets.symmetric(
-              horizontal: 12,
-              vertical: 10,
-            ),
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(8),
-              borderSide: const BorderSide(color: Colors.grey),
-            ),
-            focusedBorder: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(8),
-              borderSide: const BorderSide(color: Color(0xFF2962FF)),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildTimeField(String label) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(label, style: const TextStyle(fontSize: 12, color: Colors.grey)),
-        const SizedBox(height: 4),
-        GestureDetector(
-          onTap: _showTimePicker,
-          child: Container(
-            height: 48,
-            padding: const EdgeInsets.symmetric(horizontal: 12),
-            alignment: Alignment.centerLeft,
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.grey),
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Text(
-              selectedTime != null
-                  ? "${selectedTime!.hourOfPeriod}시 ${selectedTime!.minute.toString().padLeft(2, '0')}분 ${selectedTime!.period == DayPeriod.am ? "오전" : "오후"}"
-                  : "",
-              style: const TextStyle(fontSize: 14),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color.fromARGB(255, 255, 255, 255),
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0,
-        title: const Text(
-          "일정 추가",
-          style: TextStyle(
-            color: Colors.black,
-            fontSize: 18,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        iconTheme: const IconThemeData(color: Colors.black),
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.notifications_none),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Center(
-          child: Container(
-            padding: const EdgeInsets.all(16),
-            decoration: BoxDecoration(
-              color: Colors.white,
-              borderRadius: BorderRadius.circular(20),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.08),
-                  blurRadius: 12,
-                  offset: const Offset(0, 4),
-                ),
-              ],
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                const Text(
-                  "8월 25일 화요일",
-                  style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                ),
-                const Divider(height: 20, thickness: 1),
-                _buildLabeledField("수업 이름"),
-                const SizedBox(height: 16),
-                _buildLabeledField("간단한 설명"),
-                const SizedBox(height: 16),
-                _buildAttendanceField(),
-                const SizedBox(height: 16),
-                _buildTimeField("시간"),
-                const SizedBox(height: 20),
-                ElevatedButton(
-                  onPressed: () {
-                    // 추가 버튼 눌렀을 때 동작 구현
-                    // _attendanceNumber, selectedTime 등 활용 가능
-                  },
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: mainColor,
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    padding: const EdgeInsets.symmetric(vertical: 14),
-                  ),
-                  child: const Text(
-                    "추가",
-                    style: TextStyle(fontSize: 16, color: Colors.white),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
